+++ src/lib/enr/create.ts
@@ -2,9 +2,9 @@ import { bytesToHex } from "../utils";
 
 import { NodeId } from "./types";
 
-export function createNodeId(buffer: Buffer): NodeId {
-  if (buffer.length !== 32) {
+export function createNodeId(bytes: Uint8Array): NodeId {
+  if (bytes.length !== 32) {
     throw new Error("NodeId must be 32 bytes in length");
   }
-  return bytesToHex(buffer);
+  return bytesToHex(bytes);
 }